add app for sync board
